Feat/ms pedidos

+++ ms-pedidos/src/main/java/isi/dan/ms/pedidos/modelo/Producto.java
@@ -2,14 +2,12 @@ package isi.dan.ms.pedidos.modelo;
 
 import java.math.BigDecimal;
 
-import org.springframework.data.annotation.Id;
-
 import lombok.Data;
 
 @Data
 public class Producto {
 
-    private Long id;
+    private Integer id;
     private String nombre;
     private String descripcion;
     private BigDecimal precio;
